Move protobuf logic to semantic/
