+++ scalahost/src/main/scala/scala/meta/internal/scalahost/ScalahostPlugin.scala
@@ -5,7 +5,6 @@ import scala.tools.nsc.Global
 import scala.tools.nsc.plugins.{Plugin, PluginComponent}
 import scala.tools.nsc.typechecker.ScalahostAnalyzer
 import scala.meta.internal.scalahost.v1.LocationOps
-import scala.meta.internal.scalahost.v1.SerializationOps
 
 class ScalahostPlugin(val global: Global)
     extends Plugin
