+++ scalahost/src/main/scala/scala/meta/internal/scalahost/v1/offline/Mirror.scala
@@ -51,7 +51,7 @@ class Mirror(classpath: String, sourcepath: String)
         else Nil
       })
       .sortBy(_.getName)
-    val databases = databaseFiles.map(f => SerializationOps.fromFile(f).get)
+    val databases = databaseFiles.map(f => Database.fromFile(f).get)
     databases.foldLeft(Database())(_ append _)
   }
