+++ scalameta/semantic/src/main/protobuf/semanticdb.proto
